Changed cv_ncbiprotein function #24
To better handle the annotations of the GeneProducts with the IDs contained in the model GeneProduct ID strings added by CarveMe, the function cv_ncbiprotein was adjusted. However, the ID string has to contain 'prot_' now before the actual ID. Additionally, a function to parse the input fastA file from CarveMe to add names and labels/ locus tags for IDs that are not valid ncbiprotein IDs was added.