Updated get_model_reacs_or_metabs in entities to be more general #43
The function get_model_reacs_or_metabs now takes a new parameter. If the model's metabolites/reactions are not from the BiGG namespace the parameter col_name can now be changed to  a user-specific string. To not break the original code, col_name is by default set to 'bigg_id'.